long-->int

+++ project-bean/src/main/java/com/cykj/bean/EmployerInfo.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class EmployerInfo {
 
-  private long employerId;
+  private int employerId;
   private String employerName;
   private String employerProfile;
   private String credit;
